requirements.txt
affine==2.4.0
anyio==3.7.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
cogeo-mosaic==6.2.0
color-operations==0.1.1
cramjam==2.6.2
exceptiongroup==1.1.2
fiona==1.9.4.post1
flask[async]==2.3.2
geojson-pydantic==0.6.3
geopandas==0.13.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
mercantile==1.2.1
morecantile==4.3.0
numexpr==2.8.4
packaging==23.1
pandas==2.0.3
pillow==10.0.0
pyparsing==3.1.0
pyproj==3.5.0
pystac==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
rangehttpserver==1.3.3
rasterio==1.3.8
rio-cogeo==4.0.1
rio-stac==0.8.0
rio-tiler==5.0.3
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
starlette-cramjam==0.3.2
supermorecado==0.1.1
titiler==0.12.0
titiler-application==0.12.0
titiler-core==0.12.0
titiler-extensions==0.12.0
titiler-mosaic==0.12.0
typing-extensions==4.7.1
tzdata==2023.3
werkzeug==2.3.6
zipp==3.16.2
uvicorn
starlette
opencv-python
tqdm
pyyaml
ultralytics
torch
kornia-rs
scikit-image
albumentations
pydantic
fastapi